clang format

+++ components/parsing/src/parsing.cpp
@@ -43,7 +43,7 @@ namespace parsing {
         std::vector<std::unique_ptr<node>> tokens;
 
         size_t str_ind = 0;
-        int str_ind_section_start = 0;
+        size_t str_ind_section_start = 0;
         auto str_ind_end = input.length() - 1;
         while (str_ind <= str_ind_end) {
